stupid fking merge

+++ DystopiaEngine/DystopiaEngine/Include/Allocator/DefaultAllocImpl.h
@@ -24,7 +24,7 @@ namespace Dystopia
 	class DefaultAlloc
 	{
 	public:
-		using MetaData_t = uint64_t;
+		using MetaData_t = uint32_t;
 
 		DefaultAlloc(void);
 		~DefaultAlloc(void) noexcept;
